Added useContext for auth token and interfaces module
